[ATLAS] fix: M10 — 4 review fixes
M10-1 — extract gather_metrics into shared lib
  Moved Metric / ReadinessReport / measure_* / gather_metrics /
  render_human / render_json from scripts/bu_readiness_check.py into
  src/services/bu_readiness.py.
  Both consumers now import from the new lib:
    - scripts/bu_readiness_check.py (CLI / cron) — slim wrapper that
      opens an asyncpg connection, calls gather_metrics, prints either
      render_human or render_json, sets the right exit code.
    - src/api/routes/bu_readiness.py (REST endpoint) — drops the
      cross-package `from scripts.bu_readiness_check import ...` and
      imports cleanly from `src.services.bu_readiness`.

M10-2 — TARGET_BU_SIZE configurable
  src/config/settings.py: new field
    TARGET_BU_SIZE: int = 50_000   (env var TARGET_BU_SIZE / target_bu_size)
  measure_coverage() reads `getattr(settings, "TARGET_BU_SIZE", 50_000)`
  so the Coverage metric's denominator can be tuned per-environment
  (e.g. lower for a sandbox, higher when the addressable market estimate
  changes). The 100% target now lives in settings, not in a hard-coded
  module constant.

M10-3 — measure_outcomes logs failures
  The `try/except: n=0` block in measure_outcomes silently swallowed
  DB errors — making a "missing table" indistinguishable from a real
  zero. Replaced with `except Exception as exc: logger.error(...)` so
  the failure is visible in logs while still returning 0/FAIL so the
  rest of the report is well-formed. Test updated to assert the
  logger.error fires on a fake DB error.

M10-4 — trajectory uses standard MoM window
  Was: numerator = rows created last 30d
       denominator = rows created BEFORE that 30d window (everything
       older), which biases lower over time as the dataset matures and
       penalises healthy young systems.
  Now: numerator = rows created in last 30 days
       denominator = rows created in the 30-60 day window
       Both windows are equal length so the ratio is comparable across
       time and matches the standard month-over-month definition. The
       SQL uses two FILTER clauses with INTERVAL '30 days' and
       INTERVAL '60 days' bounds. Detail string updated to read
       "X created last 30d / Y created 30-60d ago".

Tests (tests/scripts/test_bu_readiness_check.py — now 13 passing)
  + test_coverage_uses_settings_target_bu_size
      Sets settings.TARGET_BU_SIZE=10,000 and confirms 4,000 rows = 40%.
  + test_outcomes_logs_when_table_missing
      Uses caplog to assert logger.error fires on a fake RuntimeError.
  + test_trajectory_uses_standard_mom_window
      Reads the new fetchrow shape (new_rows, prev_rows) and asserts
      detail string contains "30d" + "30-60d ago".
  Existing 11 tests carried forward unchanged (now sourced from
  src.services.bu_readiness instead of the importlib-loaded script).

Live smoke against current Supabase (sanity):
  coverage    17.29%  FAIL  (8,643 / 50,000)
  verified    15.42%  FAIL  (39 / 253)
  outcomes     0.00   FAIL  (0 cis_outreach_outcomes)
  trajectory  54.39%  PASS  (3,045 last 30d / 5,598 30-60d ago)
  Overall:    FAIL

Gates:
  ruff check  — All checks passed!  (5 .py files)
  pytest      — 13 passed in 1.19s
  Live import — `from src.api.routes.bu_readiness import router` +
                `from src.services.bu_readiness import gather_metrics` OK
